requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.30.6

SQLAlchemy==2.0.36
psycopg[binary]==3.2.3
pgvector==0.3.6

pydantic==2.10.3
python-dateutil==2.9.0.post0
requests==2.32.3

httpx==0.27.2
python-dateutil==2.9.0.post0